Creates a CWP (FAO Coordinating Working Party) grid spatial object

Creates a CWP grid spatial object.

create_cwp_grid(
  resolution = "1deg_x_1deg",
  longitude_min = -180,
  latitude_min = -90,
  longitude_max = 180,
  latitude_max = 90,
  parallel = FALSE,
  ...
)

Arguments

resolution
character expected. A string matching one of the accepted resolution values. Accepted resolutions values are "10min_x_10min", "20min_x_20min", "30min_x_30min", "30min_x_1deg", "1deg_x_1deg", "5deg_x_5deg", "10deg_x_10deg", "20deg_x_20deg" and "30deg_x_30deg".

longitude_min
integer expected. Longitude minimum value of the output grid in decimal degree. By default -180.

latitude_min
integer expected. Latitude minimum value of the output grid in decimal degree. By default -90.

longitude_max
integer expected. Longitude maximum value of the output grid in decimal degree. By default 180.

latitude_max
integer expected. Latitude maximum value of the output grid in decimal degree. By default 90.

parallel
logical expected. Run in parallel (if you have at least two processor cores). By default FALSE.

...
Others parallel options

Value

Create an object of class "SpatialPolygonsDataFrame" named grid_cwp_resolution (where resolution is the function argument).

References

CWP Handbook - http://www.fao.org/fishery/cwp/handbook/G/en From a function developed by Emmanuel Blondel emmanuel.blondel1@gmail.com.