Conversion function for latitude, longitude and cwp item

Conversion function for latitude, longitude and cwp item.

lat_lon_cwp_manipulation(
  manipulation_process,
  data_longitude = NULL,
  data_latitude = NULL,
  data_cwp = NULL,
  referential_grid_file_path,
  input_cwp_format = NULL,
  output_degree_cwp_parameter = NULL,
  output_degree_format = NULL,
  input_degree_format = NULL,
  epsg_code = as.integer(4326),
  output_cwp_format = NULL
)

Arguments

manipulation_process

character expected. A string matching one of the accepted processes. Available processes are "cwp_to_lat_lon" or "lat_lon_to_cwp".

data_longitude

character expected. Mandatory argument for "lat_lon_to_cwp" process. A string vector with longitude value(s). Formats allowed is decimal degree (with point separator) or in degree minute seconde format (xxdxx'xx'').

data_latitude

character expected. Mandatory argument for "lat_lon_to_cwp" process. A string vector with latitude value(s). Formats allowed is decimal degree (with point separator) or in degree minute seconde format (xxdxx'xx'').

data_cwp

character expected. Mandatory argument for "cwp_to_lat_lon" process. A string vector with cwp value(s). Mandatory argument for "cwp_to_lat_lon" process.

referential_grid_file_path

character expected. File path of the referential grid shape. File with .RData extension file expected.

input_cwp_format

character expected. Mandatory argument for "cwp_to_lat_lon" process. Cwp construction process related to the "centroid" or the "corner" of the square.

output_degree_cwp_parameter

character expected. Mandatory for "cwp_to_lat_lon" process. A string matching one of the accepted cwp parameter. Available parameters are "centroid" (for output coordinates in relation to cwp centroid) and "corner" (for output coordinates in relation to cwp corner).

output_degree_format

character expected. Mandatory for "cwp_to_lat_lon" process. A string matching one of the accepted output degree format. Available formats are "degree_minute_seconde" and "decimal_degree".

input_degree_format

character expected. Mandatory for "lat_lon_to_cwp" process. A string matching one of the accepted output degree format. Available formats are "degree_minute_seconde" and "decimal_degree".

epsg_code

integer expected. Mandatory for "lat_lon_to_cwp" process. An integer (on 4 digits) of one EPSG spatial reference systems related to latitude and longitude coordinates provide. Check this web site for more informations: https://www.spatialreference.org. By default 4326.

output_cwp_format

integer expected. Output format of cwp. So far, you can choose between theses formats: "centroid_11" (based on the square's centroid with a cwp on 11 characters) or "corner_11" (based on the square's corner with a cwp on 11 characters)

Value

Return a data.frame.