Consistent spatial marine area overlay

Consistent spatial marine area overlay for points, grids and polygons (related to fao, eez or ices areas).

marine_area_overlay(
  data,
  overlay_expected,
  longitude_name,
  latitude_name,
  fao_area_file_path = NULL,
  fao_overlay_level = NULL,
  auto_selection_fao = FALSE,
  eez_area_file_path = NULL,
  for_fdi_use = NULL,
  ices_area_file_path = NULL,
  silent = FALSE
)

Arguments

data

data.frame expected. R dataframe, with at least two columns with longitude and latitude values. Be careful! Your longitude and latitude data have to be in the WGS84 projection and coordinates in decimal degrees.

overlay_expected

character expected. Type of overlay output. You can choose between "fao_area", "eez_area", "fao_eez_area", "ices_area" or "all".

longitude_name

character expected. Longitude column name in your data.

latitude_name

character expected. Latitude column name in your data.

fao_area_file_path

character expected. File path of the fao area shape. You can provide a .shp or a .Rdata/.RData file.

fao_overlay_level

character expected. Level of fao accuarcy that you want for classified your data. By default, major fao fishing area are selected. Check the section details below.

auto_selection_fao

logical expected. Add a new column in the output with the most detailed overlay level available.

eez_area_file_path

character expected. File path of the eez area shape. You can provide a .shp or a .Rdata/.RData file.

for_fdi_use

logical expected. Add a new column in the output with the FDI variable "eez_indicator".

ices_area_file_path

character expected. File path of the ices area shape. You can provide a .shp or a .Rdata/.RData file.

silent

logical expected. Display or not warning information regarding projection of your spatial coordinates. By default FALSE.

Value

The function return your input data frame with one or several columns (regarding arguments' specification) which contains area classification.

Details

For the argument "fao_overlay_level", you can choose between 5 modalities (descending size classification):

ocean: ocean area

major: major fao fishing area

subarea: sub fao fishing area

division: division fao fishing area

subdivision: sub-division fao fishing area

subunit: sub-unit fao fishing area

Specificity for fao fishing area parameters: all the items above your specification (thus contain it at higher levels) will be added in the output. For example, if you select "subarea", you will also have the information about the major area concerning. If you want more information visit http://www.fao.org/fishery/area/search/en