Consistent spatial marine area overlay (related to fao area)

Consistent spatial marine area overlay (related to fao area) for points, grids and polygons.

marine_area_overlay(
  data,
  overlay_level = "major",
  longitude_name,
  latitude_name,
  tolerance = 0
)

Arguments

data
data.frame expected. R dataframe, with at least two columns with longitude and latitude values. Be careful! Your longitude and latitude data have to be in the WGS84 projection and coordinates in decimal degrees.

overlay_level
character expected. Level of accuarcy that you want for classified your data. By default, major fao fishing area are selected. Check the section details below.

longitude_name
character expected. Longitude column name in your data.

latitude_name
character expected. Latitude column name in your data.

tolerance
numeric expected. Tolerance of maximum distance between coordinates and area selected (in km). By default no tolerance (0 km).

Value

The function return your input data frame with one or several columns (regarding specification in the argument "overlay_level") which contains area classification. For avoid conflicts, new colums ended by _MAO (for marine area overlay).

Details

For the argument "overlay_level", you can choose between 5 modalities (descending size classification):

ocean: ocean area

major: major fao fishing area

subarea: sub fao fishing area

division: division fao fishing area

subdivision: sub-division fao fishing area

subunit: sub-unit fao fishing area

Specificity for fao fishing area parameters: all the items above your specification (thus contain it at higher levels) will be added in the output. For example, if you select "subarea", you will also have the information about the major area concerning. If you want more information visit http://www.fao.org/fishery/area/search/en

Examples